Check all episode files have been downloaded
Checks for the presence of all files for an episode, not just the presence of the episode folder. If network connectivity failed during download the original script would generate folders for all episodes with no files in, and then the downloaded? check would be true for all as only the directory was checked. Now the contents are checked and all files for the episode downloaded if any are missing.